Refactor neural net to follow different implem.
This refactors the neural net so that the feed forward
part follows the implementation that Morten uses. This
will probably make things easier down the road.

Specifically this makes it so that we take in an n x p
design matrix, where p are our features and n are the number
of input observations. This mirrors the linear regression design
matrix. Now we do the matrix multiplication as X @ W instead
of W @ X, which also could make functions that take assume
input @ parameters work with both linear regression and
neural nets.