Update back prop
This is pretty close to implementing backprop completely. The output
layer is correctly updating weights. The other layers are not working
quite correctly. It can successfully find the correct weights and
reach the correct prediction if it is possible to do so using only
weights from the output layer (linreg). It does not work with batches
and gradient accumulation.